precompute default skin head
